+++ src/main/java/com/surit/fixer/request/model/dto/RepairRequestDTO.java
@@ -19,7 +19,7 @@ public class RepairRequestDTO {
 
 	// ---- REPAIR_REQUESTS 컬럼 ----
 	private Long      requestId;
-	private Integer   userNo;          // 의뢰한 고객
+	private Long      userNo;          // 의뢰한 고객
 	private String    categoryCode;
 	private String    title;
 	private String    content;
@@ -31,7 +31,7 @@ public class RepairRequestDTO {
 	private String  categoryName;      // COMMON_CODE.CODE_NAME
 	private String  statusName;        // COMMON_CODE.CODE_NAME
 	private String  customerName;      // USERS.NAME
-	private Integer estimateCount;     // 이 접수에 달린 견적 수
+	private Long    estimateCount;     // 이 접수에 달린 견적 수
 
 	/**
 	 * 내가 이미 이 접수에 견적을 냈으면 그 ESTIMATE_ID, 아니면 null.
