+++ src/main/java/com/surit/fixer/verify/model/dto/FixerVerifyRequest.java
@@ -18,7 +18,7 @@ public class FixerVerifyRequest {
 
 	// → FIXER_PROFILE
 	private String  intro;
-	private Integer careerYears;
+	private Long    careerYears;
 
 	// → FIXER_PROFILE.FIXER_PHOTO_URL (고객 확인용 사진, 필수 1장)
 	private MultipartFile photoFile;
